requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2
numpy==1.26.4
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
pytest==8.3.3
bandit==1.7.9
pip-audit==2.7.3
httpx==0.27.2
